refactor: time-record refactoring
