+++ client/src/components/study/total_time/total_time.module.css
@@ -8,7 +8,3 @@
   font-weight: bold;
   border-radius: 0 0 12px 12px;
 }
-
-.button {
-  color: aliceblue;
-}
